Detect dlicked OIDs
With a clicked-event handler on the SVG we find the clicked SVG element
and extract the OID. Currently there are two options to do this: Using
dedicated attributes on the pattern declaration, or string-magic on the
xml-id of the pattern. I'm still undecided which road to follow
eventually as both approaches have pros and cons.
The actual hit-test on the SVG is also not perfect: Only the rectangular
bounding is analysed. But the relevant OID areas usually have more
complex shapes. Here, too, some additional work is required.